% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvenienceFunctions.R
\name{completeFun}
\alias{completeFun}
\title{Drop NAs by Columns}
\usage{
completeFun(data, desiredCols)
}
\arguments{
\item{data}{data.frame object of variation}

\item{desiredCols}{list of columns from which incomplete cases should be 
dropped}
}
\value{
dataframe with removed observations
}
\description{
Remove NAs based on specified columns in the data
}
